// api/models/carCreateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CarCreateRequestStatus } from './carCreateRequestStatus';
import type { CarConfigDto } from './carConfigDto';

export interface CarCreateRequest {
  /** @minLength 1 */
  name: string;
  carNumber: number;
  imageUrl?: string;
  status: CarCreateRequestStatus;
  /** @minLength 1 */
  location: string;
  inResearch: boolean;
  configuration: CarConfigDto;
}