// api/models/carCreateRequestStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CarCreateRequestStatus = typeof CarCreateRequestStatus[keyof typeof CarCreateRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarCreateRequestStatus = {
  ALLOWED: 'ALLOWED',
  BANNED: 'BANNED',
  TEMPORARY_BANNED: 'TEMPORARY_BANNED',
  TERMINATED: 'TERMINATED',
} as const;